add additional tests to CheckableInput
