+++ src/components/CheckableInput/CheckableInput.test.tsx
@@ -1,4 +1,4 @@
-import { shallow } from "enzyme";
+import { mount, shallow } from "enzyme";
 import React from "react";
 
 import CheckableInput from "./CheckableInput";
@@ -9,5 +9,41 @@ describe("CheckableInput ", () => {
       <CheckableInput type="checkbox" label="Test label"></CheckableInput>
     );
     expect(wrapper).toMatchSnapshot();
+  });
+
+  it("can disable an input", () => {
+    const wrapper = shallow(
+      <CheckableInput
+        type="checkbox"
+        label="Test disabled"
+        disabled
+      ></CheckableInput>
+    );
+    expect(wrapper.find("input[disabled]").length).toBe(1);
+  });
+
+  it("can apply the is-required class", () => {
+    const wrapper = shallow(
+      <CheckableInput
+        type="radio"
+        label="Test required"
+        required
+      ></CheckableInput>
+    );
+    expect(wrapper.find(".p-radio.is-required").length).toBe(1);
+  });
+
+  it("can set a checkbox to indeterminate state", () => {
+    const container = document.createElement("div");
+    document.body.appendChild(container);
+    const wrapper = mount(
+      <CheckableInput
+        type="checkbox"
+        label="Test label"
+        indeterminate
+      ></CheckableInput>,
+      { attachTo: container }
+    );
+    expect(wrapper.find("input").getDOMNode().indeterminate).toBe(true);
   });
 });
